Move flags and metadata over
